+++ Snake/src/GameObjects/SnakeHead.cpp
@@ -7,10 +7,10 @@ SnakeHead::SnakeHead(const sf::Vector2f& pos) {
 }
 
 void SnakeHead::Update(const float& dt) {
-	int32 maxSnakePosX = Game::Instance()->GetWindow().getSize().x + m_SnakePart->getRadius();
-	int32 minSnakePosX = -m_SnakePart->getRadius();
-	int32 maxSnakePosY = Game::Instance()->GetWindow().getSize().y + m_SnakePart->getRadius();
-	int32 minSnakePosY = -m_SnakePart->getRadius();
+	float maxSnakePosX = Game::Instance()->GetWindow().getSize().x + m_SnakePart->getRadius();
+	float minSnakePosX = -m_SnakePart->getRadius();
+	float maxSnakePosY = Game::Instance()->GetWindow().getSize().y + m_SnakePart->getRadius();
+	float minSnakePosY = -m_SnakePart->getRadius();
 
 	// Wrap the snake around the x position
 	if (m_SnakePart->getPosition().x > maxSnakePosX) {
